moving TestCreateAddIssuanceSignAndBroadcast to pset_test.go
